Wrote jest test for api endpoints
